Fixed warnings in hostpci and fixed definitions to use constant-width types where constant-width types are intended. Rationale:  BAR sizes will always be 64-bit on any platform that supports PCI. The same goes  for the return type of conf_{read,write} (32-bit) and bar_base (64-bit).
Sugar coating:
 - _Zero-effort_ upgrade path to 64-bit. 64-bit just works™. Why make it harder
   than it has to be?
 - uint64 is way shorter than unsigned long long.
 - We don't make assumptions on the compiler that can _trivially_ be avoided.

+++ vancouver/include/host/hostvf.h
@@ -32,7 +32,7 @@ public:
   /**
    * Return the base and size of a VF BAR (inside a SR-IOV capability).
    */
-  unsigned long long vf_bar_base_size(unsigned bdf, unsigned vf_no, unsigned no, unsigned long long &size, bool *is64bit=0) {
+  uint64 vf_bar_base_size(unsigned bdf, unsigned vf_no, unsigned no, uint64 &size, bool *is64bit=0) {
 
     unsigned sriov_cap = find_extended_cap(bdf, EXTCAP_SRIOV);
     if (!sriov_cap) return -1;
@@ -43,7 +43,7 @@ public:
   /**
    * Read all vf bars.
    */
-  void read_all_vf_bars(unsigned bdf, unsigned vf_no, unsigned long long *base, unsigned long long *size) {
+  void read_all_vf_bars(unsigned bdf, unsigned vf_no, uint64 *base, uint64 *size) {
 
     memset(base, 0, MAX_BAR*sizeof(*base));
     memset(size, 0, MAX_BAR*sizeof(*size));
@@ -55,7 +55,6 @@ public:
       if (is64bit) i++;
     }
   }
-
 
   /** Compute BDF of a particular VF. */
   unsigned vf_bdf(unsigned parent_bdf, unsigned vf_no)
@@ -69,7 +68,7 @@ public:
     return parent_bdf + vf_stride*vf_no + vf_offset;
   }
 
-  unsigned vf_device_id(unsigned parent_bdf)
+  uint32 vf_device_id(unsigned parent_bdf)
   {
     unsigned sriov_cap = find_extended_cap(parent_bdf, EXTCAP_SRIOV);
     if (!sriov_cap) return 0;
